Multi-Cluster example: use specific version of yq in setup.sh
The setup.sh script from the examples/multi-cluster folder needs
the 'yq` utility. However, it needs a specific version else the
script may fail to run. That version may not be the one installed
by the distribution package manager (or pip3). This commit modifies
the setup.sh script to install the version it wants locally so as
to not conflict with any other version already installed in the PATH.

Fixes: #64

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>