Converge Logical Cloud AppContext ID
This commit removes the redundancy of having the AppContext ID for
Logical Clouds (DCM) stored in both the Logical Cloud top-level DB
resources and also the Logical Cloud's StateInfo.

This is achieved by replacing all uses of the
tagContext/logicalCloudContext tag with state/StateInfo.

Plus, when it comes to fetching logical cloud information given an
AppContext ID, the CompositeAppMeta struct (in appcontext.go) has been
extended to include a LogicalCloud attribute. Now, when DCM needs to
get the logical cloud when it only has an AppContext ID, it no longer
searches the database with the tagContext (which was removed) but,
instead, fetches the CompositeAppMeta from the AppContext
(this happens during the DCM gRPC workflow).

Also making the transition to the Instantiated state take place right
after the user requests instantiation. This is what's expected of an
EMCO StateInfo (the user-intended state). Furthermore, this allows to
more accurately represent the DCM state machine and guarantee
access to the AppContext ID when needed.

Finally, DCM is now explicitly passing Namespace: "default", Level: "0"
when creating the CompositeAppMeta (for L1 LCs). In the future, access
to administrative kubeconfigs (along with selection of namespace)
should be formalized outside of both DCM and rsync.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>