Alternate K8s Rsync Plugin
In the existing K8s plugin each resource is written
to the cluster individually. The disadvantage is
that if any resource fails then the application is
in bad state on the cluster with some resources
already applied on the cluster. With this approach
either all the resources or none of the resources
are applied to the cluster

In this plugin all application resources are collected
in a temporary file, and then applied together.
More Disk space is required in this approach.

Signed-off-by: Ritu Sood <ritu.sood@intel.com>