Clean up status query parameter and status output
This commit adds a new "status" parameter to the api
that takes the values of "deployed" or "ready".  This
will replace the existing "type" parameter

type=resync == status=deployed
type=cluster == status=ready

Additionally, the top level "status" output is renamed
to "deployedStatus" and a new "readyStatus" is added at the
top level.

At the top level, the summary count objects are renamed:
"rsyncStatus" becomes "deployedCounts"
"clusterStatus" becomes "readyCounts"

At the cluster level, the "readyStatus" is renamed "connectivity"
(to avoid confusion with other uses of "readyStatus")

At the resource level, "rsyncStatus" is renamed "deployedStatus"
and "clusterStatus" is renamed "readyStatus"

Also, pick the correct status appcontext ID when an instanceId
parameter is provided to the query.

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>