Fix status query for updates and rollback
The structure of the appcontext for a deployment
intent group becomes more complex after updates and
rollback operations have been performed.

This commit updates the status handling functions
to take this into account so that the set of
apps/clusters/resources being reported on is accurate.

Also, for cluster queries (i.e. results from monitor),
the set of resources that are expected to be present is
collected from teh appcontext and then any resources
not found in the data receieved from the 'monitor' process
are give a status of 'Not Present'.

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>