Refactor single-cluster
Refactor single-cluster (mostly setup.sh) to allow for the choice of
logical cloud level. This change may be expanded to other examples in
the future, but for now only single-cluster.

The selection of the logical cloud level, as well as existing variables
such as KUBE_PATH/HOST_IP, and what ports the EMCO microservices are
running on, are now all specified in a configuration file called "config"
inside the single-cluster folder, although environment variables still
work for the existing HOST_IP and KUBE_PATH and the new LOGICAL_CLOUD_LEVEL.

The default config file is ready for a
typical EMCO K8s installation, while config.local is pre-configured with
the default ports where the EMCO binaries run when running without K8s.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>