Parameterize Helm version
The Helm version to use for the purpose of building the build base
image, among other future uses, is now defined in config.txt, where
other variables are also defined such as the version of Go.

When changing this version, the version of build base itself should also
be bumped.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>