Handle API errors for an interdependent controller
The error handling for some API calls is not working as expected.
If a controller makes an internal call to another controller while processing the request,
and if the dependent call fails, it returns an internal server error to the client.

For example, in the case of orchestrator returns a deployment intent group not found error to DTC.
This error is defined only in the orchestrator. The current API error handling logic matches this error with the error listed in DTC.
Since DTC does not have this error listed, it will return an internal server error to the client.

This change captures all the shared errors and returns the appropriate error to the client.

Signed-off-by: Subin John <subin.john@intel.com>