Split L1LC cluster-specific code into own function
This splits off the code that was specific to the provisioning of
DCM L1 LC cluster references in appcontext, into its own function.

This function can then later be reused by other parts of the code that
need to provision clusters references in appcontext (outside of the
normal DCM Instantiate).

This also fixes an effectless bug where the "logical-cloud" appcontext
app data was being over and over rewritten in appcontext, once per
cluster reference.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>