Make rsync update when Logical Cloud updates
If a Logical Cloud update is issued, and that LC happens to already be
instantiated, a new appcontext is built and an update call to rsync is
sent, requesting changes to be calculated and clusters to be
updated/instantiated as needed in order to bring the current state of
the DCM logical cloud to reality.

This commit also modifies the cluster add/remove logic to allow for the
list of instantiated clusters of a logical cloud to be updated without
terminating the logical cloud.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>