Separate PUT method from the /update endpoint
Both are types of "update" operations as it relates to a Logical Cloud.
However, a PUT on the LogicalCloud resource itself is simply for
updating that LogicalCloud resource's attributes. Whereas a POST call to
the /update API method of that LogicalCloud represents a request to
update the instantiation of that Logical Cloud, which involves rsync.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>