Correctly use LC private key for all CSRs created
This fixes a critical issue where a new private key was generated each
time a new CSR had to be generated for a given Logical Cloud, even
though only the very first private key associated with that Logical
Cloud would remain stored in the database. This meant that any cluster
reference after the first one (whether created before logical cloud
instantiation or using /update) would actually be given an unusable L1
kubeconfig file.

As part of the above, the step of generating a private key has been
moved from the instantiation of the logical cloud to its creation.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>