Add clarifications regarding LC User Permissions
In the design/Logical_Clouds.md, clarified in 2 different places that
Standard/Privileged logical clouds require **at least 1 User Permission
must be created** for this level of Logical Cloud, **with the namespace
set to the namespace of the respective Logical Cloud** itself,
otherwise the instantiation will fail.

Also slightly modified the sections where the clarifications were made
in order to align to where the examples are now. However, other sections
remain untouched and as such are still outdated since the logical cloud
examples got removed from the emcoctl folder.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>