Multiple improvements to reduce code duplication
This commit is focused on reducing code duplication between DCM and the
Orchestrator. It's not quite enough to remove the cross-dependency
entirely, but it brings duplication to a minimum.

Summary of changes:
- Sync up cluster.go and logicalcloud.go to latest versions from DCM (in
the Orchestrator)
- Removed unnecessary GetAppContextStatus. This reduces the
depth of dependencies between orchestrator and DCM, and reuses existing
code. It also happens to reduce a few CPU cyles.
- Remove duplicate Metadata type definitions that were defined for
Cluster References and Logical Clouds in DCM. Reuse the type Metadata
from orchestrator types.go instead.
- Move non-DB cluster.go and logicalcloud.go code/types into a new package
common in the Orchestrator.
- Trim the remaining cluster.go and logicalcloud.go copies in the
Orchestrator to the bare minimum necessary to support the dependency of
Orchestrator on DCM (total 3 methods).
- Adapt DCM unit tests.
- vscode also fixed a bunch of the code (formatting-wise).

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>