Top-level README Installation guide for EMCO
Add a whole new section to the top-level README file for EMCO about how
to install EMCO with only the public Helm charts available (the
recommended installation procedure for EMCO first-timers).

And move most of the existing contentinto docs/BUILD.md where more
advanced methods of building and deploying EMCO are presented (a link
is added to this documentation file in the top-level README file).

Minor Helm chart corrections.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>