Bump Helm pkg versions to include new controllers
This commit bumps emco and emco-services helm packages to 1.0.1, since
the charts have been modified to include the new controllers: swc, tac
and ca-certs, ahead of the 22.06 release.

This commit also updated the RELEASE-GITLAB.md instructions as well as
fix a couple Helm issues to sneaked in via the new controllers.

Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>