Add placeholder context.Background() to services
This has no functional changes. Previously a context.Background() was
created at the outgoing request; this change only pushes that up the
call stack.

It is intended that each service will pass the incoming context
through to outgoing requests, but for now the placeholder
context.Background() is inserted in order to compile against the
changes in the orchestrator and rsync modules.

Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>